antigravity-youtube/requirements.txt
yt-dlp
youtube-transcript-api
pyyaml
requests
